requirements.txt
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
joblib==1.3.2
jsonlines==4.0.0
lightning-utilities==0.9.0
lit==17.0.6
matplotlib==3.7.4
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
pandas==2.0.3
pathtools==0.1.2
pytorch-lightning==2.0.8
pytz==2023.4
PyYAML==6.0.1
referencing==0.35.1
regex==2023.8.8
rich==13.5.2
rpds-py==0.18.1
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.1
sentry-sdk==1.40.0
setproctitle==1.3.3
submitit==1.5.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.1.2
torchvision==0.15.2
tqdm==4.66.1
wandb==0.15.10